create ObserverUtils
